wip migration to macroquad

+++ src/position.rs
@@ -8,10 +8,10 @@ pub struct Position {
 }
 
 impl Position {
-    pub fn interpolate_x(&self, alpha: f32) -> i32 {
-        ((self.x * alpha) + (self.prev_x * (1.0 - alpha))).round() as i32
+    pub fn interpolate_x(&self, alpha: f32) -> f32 {
+        ((self.x * alpha) + (self.prev_x * (1.0 - alpha))).round()
     }
-    pub fn interpolate_y(&self, alpha: f32) -> i32 {
-        ((self.y * alpha) + (self.prev_y * (1.0 - alpha))).round() as i32
+    pub fn interpolate_y(&self, alpha: f32) -> f32 {
+        ((self.y * alpha) + (self.prev_y * (1.0 - alpha))).round()
     }
 }
